tiktok_bot/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
decorator==5.1.1
distro==1.9.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
jiter==0.8.2
moviepy==2.1.2
numpy==2.2.3
openai==1.63.0
outcome==1.3.0.post0
pillow==10.4.0
playwright==1.50.0
proglog==0.1.10
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pyee==12.1.1
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
schedule==1.2.2
selenium==4.28.1
setuptools==65.5.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.3.0
websocket-client==1.8.0
wsproto==1.2.0
